Refactoring customer info
Customer info moved into it's own component
