+++ one_click/client/components/OneClickLayout/components/CustomerInfo/index.js
@@ -1,0 +1,1 @@
+export { default as CustomerInfo } from './CustomerInfo'
